fizzbuzz.html
This is basically a file with a script that displays a set  of number from 1 to 100. With those divisible by 3, console.log will be used to print "Fizz" instead of the number. When the number is divisible by 5, console.log will be used to print"Buzz". If the number is divisible by both 3 and 5, the display will be "FizzBuzz" (and still print those  numbers divisible by 3 and 5 separately).